+++ packages/atlas/test/atlas_test.dart
@@ -51,6 +51,8 @@ main() {
         polylines: Set<Polyline>(),
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -71,6 +73,8 @@ main() {
           polylines: Set<Polyline>(),
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -99,6 +103,8 @@ main() {
         polylines: Set<Polyline>(),
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -120,6 +126,8 @@ main() {
           polylines: Set<Polyline>(),
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -135,6 +143,8 @@ main() {
         polylines: Set<Polyline>(),
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -155,6 +165,8 @@ main() {
           polylines: Set<Polyline>(),
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -181,6 +193,8 @@ main() {
         circles: circles,
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -202,6 +216,8 @@ main() {
           circles: circles,
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -240,6 +256,8 @@ main() {
         circles: Set<Circle>(),
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -261,6 +279,8 @@ main() {
           circles: Set<Circle>(),
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -291,6 +311,8 @@ main() {
         circles: Set<Circle>(),
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -312,6 +334,8 @@ main() {
           circles: Set<Circle>(),
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -331,6 +355,8 @@ main() {
         onTap: onTap,
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -353,6 +379,8 @@ main() {
           onTap: onTap,
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -372,6 +400,8 @@ main() {
         onLongPress: onLongPress,
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -394,6 +424,8 @@ main() {
           onLongPress: onLongPress,
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -409,6 +441,8 @@ main() {
         polylines: Set<Polyline>(),
         showMyLocation: true,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -430,6 +464,8 @@ main() {
           polylines: Set<Polyline>(),
           showMyLocation: true,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -445,6 +481,8 @@ main() {
         polylines: Set<Polyline>(),
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -465,6 +503,8 @@ main() {
           polylines: Set<Polyline>(),
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -480,6 +520,8 @@ main() {
         polylines: Set<Polyline>(),
         showMyLocation: false,
         showMyLocationButton: true,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -502,6 +544,8 @@ main() {
           polylines: Set<Polyline>(),
           showMyLocation: false,
           showMyLocationButton: true,
+          mapType: MapType.normal,
+          showTraffic: false,
         ),
       ).called(1);
     });
@@ -517,6 +561,8 @@ main() {
         polylines: Set<Polyline>(),
         showMyLocation: false,
         showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
       )).thenReturn(Container(key: mapKey));
       await tester.pumpWidget(
         MaterialApp(
@@ -537,6 +583,87 @@ main() {
           polylines: Set<Polyline>(),
           showMyLocation: false,
           showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
+        ),
+      ).called(1);
+    });
+
+    testWidgets(
+        'should call provider build method with correct arguments when mapType is supplied',
+        (WidgetTester tester) async {
+      when(provider.build(
+        initialCameraPosition: initialCameraPosition,
+        markers: Set<Marker>(),
+        circles: Set<Circle>(),
+        polygons: Set<Polygon>(),
+        polylines: Set<Polyline>(),
+        showMyLocation: false,
+        showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: false,
+      )).thenReturn(Container(key: mapKey));
+      await tester.pumpWidget(
+        MaterialApp(
+          home: Scaffold(
+            body: Atlas(
+              initialCameraPosition: initialCameraPosition,
+            ),
+          ),
+        ),
+      );
+      expect(find.byKey(mapKey), findsOneWidget);
+      verify(
+        provider.build(
+          initialCameraPosition: initialCameraPosition,
+          markers: Set<Marker>(),
+          circles: Set<Circle>(),
+          polygons: Set<Polygon>(),
+          polylines: Set<Polyline>(),
+          showMyLocation: false,
+          showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: false,
+        ),
+      ).called(1);
+    });
+
+    testWidgets(
+        'should call provider build method with correct arguments when showTraffic is enabled',
+        (WidgetTester tester) async {
+      when(provider.build(
+        initialCameraPosition: initialCameraPosition,
+        markers: Set<Marker>(),
+        circles: Set<Circle>(),
+        polygons: Set<Polygon>(),
+        polylines: Set<Polyline>(),
+        showMyLocation: false,
+        showMyLocationButton: false,
+        mapType: MapType.normal,
+        showTraffic: true,
+      )).thenReturn(Container(key: mapKey));
+      await tester.pumpWidget(
+        MaterialApp(
+          home: Scaffold(
+            body: Atlas(
+              initialCameraPosition: initialCameraPosition,
+              showTraffic: true,
+            ),
+          ),
+        ),
+      );
+      expect(find.byKey(mapKey), findsOneWidget);
+      verify(
+        provider.build(
+          initialCameraPosition: initialCameraPosition,
+          markers: Set<Marker>(),
+          circles: Set<Circle>(),
+          polygons: Set<Polygon>(),
+          polylines: Set<Polyline>(),
+          showMyLocation: false,
+          showMyLocationButton: false,
+          mapType: MapType.normal,
+          showTraffic: true,
         ),
       ).called(1);
     });